+++ examples/data-sources/nxos_subinterface/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_subinterface" "example" {
+  interface_id = "eth1/10.124"
+}