+++ examples/resources/nxos_subinterface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_subinterface.example "sys/intf/encrtd-[eth1/10.124]"